# Conflicts:
#	dependencies/dependencies.gradle

+++ core/src/main/java/com/exozet/android/core/utils/StringExtensions.java
@@ -22,7 +22,7 @@ final public class StringExtensions {
         if (isEmpty(s))
             return "";
 
-        return s.substring(0, (int) MathExtensions.clamp(s.length(), 0, maxLength));
+        return s.substring(0, MathExtensions.clamp(s.length(), 0, maxLength));
     }
 
     /**
